Fix compile time warnings when building for the CSKY target on a 32-bit host.
incldue	* opcode/csky.h (CSKY_ISA_FLOAT_7E60): Use a long long type for
	this value.

opcodes	* csky-dis.c (csky_output_operand): Coerce the immediate values to
	long before printing.

+++ opcodes/csky-dis.c
@@ -675,7 +675,7 @@ csky_output_operand (char *str, struct operand const *oprnd,
 	  }
 	double d = 0;
 	memcpy (&d, &dvalue, sizeof (double));
-	sprintf (buf, "%lf\t// imm9:%4ld, imm4:%2ld", d, imm8, imm4);
+	sprintf (buf, "%lf\t// imm9:%4ld, imm4:%2ld", d, (long) imm8, (long) imm4);
 	strcat (str, buf);
 	break;
       }
